Addressing some PR comments. Added feature flag.
